+++ CHANGES.rst
@@ -11,7 +11,11 @@ Note: updated docs now found at `<http://refl1d.readthedocs.org>`_
 * adjust dQ with sample broadening rather than calculating it from dL and dT
 * fix json output so saving the fit doesn't raise and error
 * fix functional magnetic profile so [rho, rho] is not treated as rho, theta.
+* fix plots to show magnetic angle when not the standard magnetic angle
+* fix legends when show magnetic angles on plots
 * wxpython is now an optional dependency
+
+**BREAKING CHANGE**: functional magnetic profile now defaults to thetaM=270
 
 2020-02-18 v0.8.10
 ==================